requirements.txt
Flask==3.1.0
SQLAlchemy==2.0.39
PyMySQL==1.1.1
python-dotenv==1.0.1
cryptography==44.0.2
